requirements.txt
#Django
Django~=3.0.6

#Databases
psycopg2-binary

# Utilities
python-dateutil
python-dotenv


